+++ plugins/src/com/oracle/oci/eclipse/ui/explorer/database/actions/CustomADBInstanceAction.java
@@ -1,0 +1,13 @@
+package com.oracle.oci.eclipse.ui.explorer.database.actions;
+
+import com.oracle.oci.eclipse.ui.explorer.common.BaseAction;
+
+public class CustomADBInstanceAction extends BaseAction {
+
+    
+    @Override
+    protected void runAction() {
+        
+    }
+
+}